requirements.txt
# Backend Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.2
aiofiles>=24.1.0

# Database - PostgreSQL with PostGIS
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23
geoalchemy2>=0.14.2
alembic>=1.13.0

# Caching - Redis
redis>=5.0.1
aioredis>=2.0.1
hiredis>=2.3.2

# Earth Engine & Geospatial
earthengine-api>=1.6.12
geemap>=0.35.0
folium>=0.18.0
shapely>=2.0.2
geopandas>=0.14.1
pyproj>=3.6.1

# Scientific Computing
numpy>=1.26.0
pandas>=2.1.3
scipy>=1.11.4
scikit-learn>=1.3.2

# Climate Analysis
xarray>=2023.12.0
netCDF4>=1.6.5
rasterio>=1.3.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

